requirements.txt
openai
supabase
psycopg2-binary
requests
python-dotenv
slack_bolt
slack_sdk 